ws/requirements.txt
tornado==6.4
pymongo==4.7.2
python-dateutil==2.9.0
PyJWT==2.8.0
bcrypt==3.2.0
